Refactor agent-browser session and result structure
