+++ extensions/agent-browser/lib/results/text.ts
@@ -1,0 +1,40 @@
+/**
+ * Purpose: Hold tiny text and ref-id helpers shared by result renderers.
+ * Responsibilities: Convert unknown values to text, count lines, normalize whitespace, truncate labels, and sort ref ids naturally.
+ * Scope: Generic pure utilities only; no agent-browser command policy belongs here.
+ * Usage: Imported by envelope, presentation, snapshot, and diagnostic helpers.
+ * Invariants/Assumptions: Helpers are deterministic and side-effect free.
+ */
+
+export function stringifyUnknown(value: unknown): string {
+	if (typeof value === "string") return value;
+	if (typeof value === "number" || typeof value === "boolean") return String(value);
+	if (value === null || value === undefined) return "";
+	try {
+		return JSON.stringify(value, null, 2);
+	} catch {
+		return String(value);
+	}
+}
+
+export function countLines(text: string): number {
+	return text.length === 0 ? 0 : text.split("\n").length;
+}
+
+export function normalizeWhitespace(text: string): string {
+	return text.replace(/\s+/g, " ").trim();
+}
+
+export function truncateText(text: string, maxChars: number): string {
+	if (text.length <= maxChars) return text;
+	return `${text.slice(0, Math.max(1, maxChars - 1))}…`;
+}
+
+export function compareRefIds(left: string, right: string): number {
+	const leftMatch = left.match(/^(?:[a-zA-Z]+)?(\d+)$/);
+	const rightMatch = right.match(/^(?:[a-zA-Z]+)?(\d+)$/);
+	if (leftMatch && rightMatch) {
+		return Number(leftMatch[1]) - Number(rightMatch[1]);
+	}
+	return left.localeCompare(right);
+}
